+++ cryptofeed/backends/postgres.py
@@ -16,18 +16,20 @@ from cryptofeed.defines import FUNDING, OPEN_INTEREST, TICKER, TRADES, LIQUIDATI
 
 
 class PostgresCallback:
-    def __init__(self, host='127.0.0.1', user=None, pw=None, db=None, table=None, numeric_type=float, **kwargs):
+    def __init__(self, host='127.0.0.1', user=None, pw=None, db=None, table=None, numeric_type=float, cache_size=0, **kwargs):
         """
-        library: str
-            Postgres library. Will be created if does not exist.
-        key: str
-            setting key lets you override the symbol name.
-            The defaults are related to the data
-            being stored, i.e. trade, funding, etc
-        kwargs:
-            if library needs to be created you can specify the
-            lib_type in the kwargs. Default is VersionStore, but you can
-            set to chunkstore with lib_type=Postgres.CHUNK_STORE
+        host: str
+            Database host address
+        user: str
+            The name of the database role used for authentication.
+        db: str
+            The name of the database to connect to.
+        pw: str
+            Password to be used for authentication, if the server requires one.
+        table: str
+            Table name to insert into. Defaults to default_table that should be specified in child class
+        cache_size: int
+            Number of lines to cache before writing the whole as a batch. Defaults to 0, meaning no caching.
         """
         self.conn = None
         self.table = table if table else self.default_table
@@ -37,17 +39,42 @@ class PostgresCallback:
         self.pw = pw
         self.host = host
 
+        self._cache_size = cache_size
+        self._cache = []
+        self._cache_counter = 0
+
     async def _connect(self):
         if self.conn is None:
             self.conn = await asyncpg.connect(user=self.user, password=self.pw, database=self.db, host=self.host)
 
     async def write(self, feed: str, pair: str, timestamp: float, receipt_timestamp: float, data: dict):
-        await self._connect()
-        async with self.conn.transaction():
-            time = dt.utcfromtimestamp(timestamp)
-            rtime = dt.utcfromtimestamp(receipt_timestamp)
+        time = dt.utcfromtimestamp(timestamp)
+        rtime = dt.utcfromtimestamp(receipt_timestamp)
 
-            await self.conn.execute(f"INSERT INTO {self.table} VALUES('{feed}','{pair}','{time}','{rtime}',{data})")
+        self._cache.append(f"('{feed}','{pair}','{time}','{rtime}',{data})")
+        self._cache_counter += 1
+
+        if self._cache_counter > self._cache_size:
+            await self.write_cache()
+
+    async def write_cache(self):
+        await self._connect()
+
+        async with self.conn.transaction():
+            args_str = ','.join(line for line in self._cache)
+
+            self._cache_counter = 0
+            self._cache.clear()
+
+            try:
+                await self.conn.execute(f"INSERT INTO {self.table} VALUES {args_str}")
+            except asyncpg.UniqueViolationError:
+                # when restarting a subscription, some exchanges will re-publish a few messages
+                pass
+
+    async def stop(self):
+        if self._cache_counter > 0:
+            await self.write_cache()
 
 
 class TradePostgres(PostgresCallback, BackendTradeCallback):