+++ CHANGES.md
@@ -9,6 +9,7 @@
   * Feature: Orderbook sequence number validation for HitBTC
   * Feature: Kraken orderbook checksum support in Kraken
   * Feature: KrakenFutures sequence number check added
+  * Feature: Add optional caching to postgres backend
 
 ### 1.6.1 (2020-11-12)
   * Feature: New kwarg for exchange feed - `snapshot_interval` - used to control number of snapshot updates sent to client